requirements.txt
pyyaml
requests==2.22.0
beautifulsoup4==4.8.1
